Added scraper.py
scraper.py queries for team stats for the current season, and will display them out to the console. Eventually this will be used to maintain an accurate record of statistics that will be used for the other features of this app (Or support only the season total stats framework, not the week to week stats framework).